requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile test.in
#
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via
    #   chalice
    #   pytest
backcall==0.2.0
    # via ipython
black==22.1.0
    # via ipython
blessed==1.19.1
    # via inquirer
botocore==1.23.53
    # via chalice
certifi==2021.10.8
    # via requests
chalice==1.26.6
    # via pytest-chalice
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via
    #   black
    #   chalice
    #   safety
coverage[toml]==6.3.1
    # via pytest-cov
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
dparse==0.5.1
    # via safety
executing==0.8.2
    # via stack-data
factory-boy==3.2.1
    # via pytest-factoryboy
faker==13.0.0
    # via factory-boy
flake8==3.8.3
    # via
    #   -r test.in
    #   pytest-flake8
freezegun==1.1.0
    # via -r test.in
idna==3.3
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via pytest-factoryboy
inquirer==2.9.1
    # via chalice
ipdb==0.13.9
    # via -r test.in
ipython==8.0.1
    # via
    #   -r test.in
    #   ipdb
jedi==0.18.1
    # via ipython
jmespath==0.10.0
    # via
    #   botocore
    #   chalice
matplotlib-inline==0.1.3
    # via ipython
mccabe==0.6.1
    # via flake8
more-itertools==8.12.0
    # via pytest
multidict==6.0.2
    # via yarl
mypy-extensions==0.4.3
    # via
    #   black
    #   chalice
packaging==21.3
    # via
    #   dparse
    #   pytest
    #   safety
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==2.5.0
    # via black
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.27
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pycodestyle==2.6.0
    # via flake8
pyflakes==2.2.0
    # via flake8
pygments==2.11.2
    # via ipython
pyparsing==3.0.7
    # via packaging
pytest==5.4.3
    # via
    #   -r test.in
    #   pytest-chalice
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
    #   pytest-flake8
    #   pytest-vcr
pytest-chalice==0.0.5
    # via -r test.in
pytest-cov==3.0.0
    # via -r test.in
pytest-django==4.5.2
    # via -r test.in
pytest-factoryboy==2.1.0
    # via -r test.in
pytest-flake8==1.0.7
    # via -r test.in
pytest-vcr==1.0.2
    # via -r test.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   faker
    #   freezegun
python-editor==1.0.4
    # via inquirer
pyyaml==6.0
    # via
    #   chalice
    #   dparse
    #   vcrpy
readchar==3.0.5
    # via inquirer
requests==2.27.1
    # via safety
safety==1.9.0
    # via -r test.in
six==1.16.0
    # via
    #   asttokens
    #   blessed
    #   chalice
    #   python-dateutil
    #   vcrpy
stack-data==0.1.4
    # via ipython
toml==0.10.2
    # via
    #   dparse
    #   ipdb
tomli==2.0.1
    # via
    #   black
    #   coverage
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.0.1
    # via black
urllib3==1.26.8
    # via
    #   botocore
    #   requests
vcrpy==4.1.1
    # via pytest-vcr
wcwidth==0.2.5
    # via
    #   blessed
    #   prompt-toolkit
    #   pytest
wheel==0.37.1
    # via chalice
wrapt==1.13.3
    # via vcrpy
yarl==1.7.2
    # via vcrpy

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
